requirements.txt
loguru
pydantic
python-dotenv
proxystr
patchright
playwright
python-socketio
flask[async]
hypercorn

pyautogui
opencv-python